pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Owais Jamil"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.0"}
psycopg = {extras = ["binary"], version = "^3.2.2"}
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.15.1"}
django-extensions = "^3.2.3"
werkzeug = "^3.0.4"
django-cors-headers = "^4.4.0"
pyjwt = "^2.9.0"
ipython = "^8.28.0"
ipdb = "^0.13.13"
pydantic = "^2.9.2"
celery = "^5.4.0"
flower = "^2.0.1"
django-celery-results = "^2.5.1"
redis = "^5.1.1"
faker = "^30.3.0"
watchdog = "^5.0.3"
pandas = "^2.2.3"
pandas-stubs = "^2.2.3.241009"
loguru = "^0.7.2"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
click = "^8.1.7"
rich = "^13.9.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
mypy = "^1.11.2"
celery-types = "^0.22.0"
coverage = "^7.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "server.settings"

[tool.coverage.report]
omit = [
    "api/services/*",
    "**/migrations/*",
    "**/__init__.py",
    "**/admin.py",
    "**/apps.py",
    "**/urls.py",
    "**/tests.py",
    "**/constants.py",
    "manage.py",
    "server/*",
    "web/*",
    "docs/*",
]
